Create and work on Index.html
First i created the index.html page and all the other pages of the site.  I then made a colour pallete to help with the site themeing. First I made the menu and the using css and a but of js I made it responsive. After that I added the dropdown menu as shown in the plan.