Making the menu sticky
First I centered the Menu bar's links and then made the menu sticky. First I had problems where the drop downcontent is hidden behind the page so I had to fix this by adding it to the js :        dropdown.classList.add("sticky").
After that the css made the width 100% so I had to make a second css class called sticky-dropdown so that the drop down's content doesnt take up the entire page width when shown. My next problem was with the mobile veiw, I tried adding in the js but that didnt work so next I changed the js to toggel the responsive class and that worked! I then had to make the button a toggel so after a quick google I decied to do a js version of it.