youtube hero video
I managed to get the video done with no errors but then I came across the fault that the menu bar is hidden by the video until you scroll downwards to the content wich is not wanted. I then added to the main css property (.Menu) z-index: 2; which fixed it up on desktop veiw. I quickly tested if it works on mobile but to my surpise the button that lets you veiw the menu is hidden so I had to fix this. While attempting to fix this I realised that no my little z index did not fix this but it was the sticky nav that did this so I read up about it and then fixed it by add position asbolute. But we still had the error of the mobile button not showing up so after a bit of trail and error i fixed it by removing some css code