Bug Fixing Continued
To fix most of the mobile bugs surprisingly all I had to do was remove the css part that hid the login button on smaller screens. This CSS element was going to be used later but I forgot about it and it turns out it was the reason for multiple errors in my code.

I then made the video loop.

After this I found another bug where the changelog would show on mobile views where it should be hidden until button press.  I fixed this by removing the video container class from the media veiw specifications and that seemed to fix it. After that I cleaned up my css and html code.

Now I noticed that the video was not found so I fixed it by removing the accidental -VC at the end of the looping parameter for the iframe video.

Now the changelog dropdown isn't working on load but will work after it becomes sticky. So I had to fix this by making the dropdown content constantly dropdown-content sticky-dropdown.

Just after I thought I was finished I realized the Q&A link was directing to upcoming features page so I modified that to be correct.