Breif description and
First I had to fix the scroll to top button not showing up by moving all the js that i previously moved back to the html file because it turns out it doesn't work externally that was just my browers loading weirdly.

After that I added custom scroll bars to better fit the themeing.

Next I began work on the brief description after adding tips to various buttons and links. For the brief I just center a <p> element and wrote directly into it which means I finally means I can remove the place holder text.

Next I went and worked on the Images and bullet point  list for the features. I decided to go with columns and boxes.

[STARTED VIDEO LOG]