Index Page Q&A [CONTINUED]
First I made the site look a tiny bit nicer by adding a box shadow to buttons, images and some other elements. Then I went and replaced the place holder content with actual answers to the questions. After that I wrote a small paragraph directing users to the q and a page if they happen to have more questions.