Footer
For the footer I decided to break off the plan and change it up to be a bit different.  First I started by making greay background and then I  added a marquee and some social links. After that I grabbed some Icons from font awesome to make it look better.

I was testing out if the footer was working in a mobile veiw I noticed that  sticky class got removed off the menu so I added it back and changed the js so that is was right in there aswell. I also notice that when trying to open the changelog dropdown menu it would take you to that page. The error was happening becuase you cant hover on mobile so to show the menu the user would have to click on the button which is also a link. My fix was to add a empty 'a href' and that did the trick. I also renamed the dropdown menu to "Changes".

A few finishing touches were to change the transition speed so that it was faster.