Make Template . html
I am going to make a template file so that creating other pages wont be as hard. This template file is going to be based off index.html and most elements are going to be copied. I did this by first coping every thing from index.html across to template.html and up dated the href's folder path. The I replaced the text with filler text ready for copy and pasting.