About.html
To make about.html I just copied the medium sized paragraph and image from the helpful template page and replaced the filler content with actual information. After that I opened the game it self and took some screenshots for images that were going to be used in the website.

After that I changed the CSS classes' names to be more relevant.