ChangeLog.html
First I made changelog.html and then copied the base code from template.html, then i noticed that the scroll to top button wasn't working so I had to fix that. The problem was that when I was messing around with the js earlier I accidently deleted the function to display the button.

Afte I fixed that I tried to make use of the changelog from the patcher but it doesnt seem to want to work so I had to slowly, row by row, copy the items into a html list.