Make Upcoming Features.html
First I copied  the elements that I needed from template. html then I painstakingly deleted some contrnt and had to do this remoltly logged in from my phone; may not do any work on this for a week idk.

After that a decide to make a tabel and  the fill in the filler content with actual infomation.