Notice.Html
First I copied the code from the Changelog  page across to this current one and then rewrote the page contents to mach.

After that I coded the notice page, which is a basic JS slide show. I managed to do this quite fast and the  small issue I had was forgeting to add a plus like this (showSlide(slideIndex += n);)