Status.html
First I (as allways) copied the contents I needed from tmeplate.html. Then I made a js service check which all that it really does is try load an Image from a page. And if that image isnt found then it changes the css class off the button and alerts the user with a dialog. Then I decided to implement cookies, with cookies the funcxtion wont have to be run on every page load but run every day. This helps page load time and helps the user expericene that the alert box wont anooy them every reload.