Login.html
NOTE LOCAL FILES MAY DIFFER FROM SERVER FILES

To make login . html I had to do a lot, like learn hashing and php. First a I used a basic tutorial to make a login system. Then I made all the files local so that I could edit the page's css.  I then designed it to be a modern mix between the plan and the current site. After that I spent AGES trying to figure out how to make a encrytion system for the password.  I did this for two reasons 1 beign the it makes the password more secure (hence implication) and 2 beign that the login plugin i use for the game is encrypted so they have to match for it to work. I spent a lot of time doining this and then relised that the sytem was hasing instead of encrypting. I then spent the rest of the day learning abt bcrypt and other hashing. After that i had to implemnt the extenral page into the site, First I  tried a fullpage ifram but i THINK that it cant store the session that way so i just redirected the user to the login site.