Animation and game launching
First I wend through and added animations to the about page. Then I had a random thought of incuding a button to open the launcher from the site. I did this by first including code in the installer to add the launcher to the registry, then I added a centered button that opens the prospect. After that i went through the other pages animating them where needed.