Change from reconops
After a bit of thought I changed the game the website was on from recon ops to sunnyland, this was because ReconOps may be classed as voilent or unaceptable for marking. First I went through the index page and changed the text content to better describe the new game, and i also changed the color theme to match the game.