Bug fixing and changing
First I fixed a lot of the JS, then I went through and a loader to every page except for the status.html page because if a site was down then the page would be loading forever. After that, I added some JS to change whether the download button downloads the game or it launches the game if a certain cookie says that the game has been played.  After that, I changed the changelog so that it would be automatically generated from the JSON that stores the changelog for the launcher. Next, I edited the q and a page to have Sunnyland contents.