Transitions, Demo, Launcher
I first added a demo page that has an embeded copy of the WebGL demo hosted on the Github page. I then added some JS that changed the download buttons to lauch buttons if the user had a cookies stating they owned the game. Finnaly I added a transition between pages that is simmalar to the one in-game between levels. This transition was made using a css clipath animation and positioning JS.